PythonBackend/requirements.txt
fastapi
uvicorn
pandas
numpy
python-dotenv
langchain
langchain-openai
langchain-chroma
chromadb